requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==1.10.13
chromadb==0.4.24
sentence-transformers==2.2.2
transformers==4.41.2
accelerate==0.30.1
pypdf==3.17.4
python-multipart==0.0.9
streamlit==1.29.0
google-generativeai==0.3.2